fix(app-page-builder): use secondary config priority for editor configs [skip ci]
