+++ packages/app-page-builder/src/blockEditor/Editor.tsx
@@ -12,13 +12,12 @@ import {
 } from "~/admin/graphql/pages";
 import createElementPlugin from "~/admin/utils/createElementPlugin";
 import { createStateInitializer } from "./createStateInitializer";
-import { DefaultEditorConfig } from "~/editor";
-import { BlockEditorConfig } from "./config/BlockEditorConfig";
 import { BlockWithContent } from "~/blockEditor/state";
 import { createElement } from "~/editor/helpers";
 import { PbEditorElement } from "~/types";
 import elementVariablePlugins from "~/blockEditor/plugins/elementVariables";
 import { usePageBlocks } from "~/admin/contexts/AdminPageBuilder/PageBlocks/usePageBlocks";
+import { DefaultBlockEditorConfig } from "~/blockEditor/config/DefaultBlockEditorConfig";
 
 export const BlockEditor = () => {
     plugins.register(elementVariablePlugins());
@@ -68,8 +67,7 @@ export const BlockEditor = () => {
 
     return (
         <React.Suspense fallback={<EditorLoadingScreen />}>
-            <DefaultEditorConfig />
-            <BlockEditorConfig />
+            <DefaultBlockEditorConfig />
             <LoadData>
                 {/* PbEditor components is a shell component, which is decorated in `src/PageBuilder.tsx`. */}
                 {/* This allows developers to override how the editor component is loaded and mounted. */}
