+++ packages/app-page-builder/src/blockEditor/config/eventActions/EventActionHandlerDecorator.tsx
@@ -11,7 +11,7 @@ import { useBlock } from "~/blockEditor/hooks/useBlock";
 
 type ProviderProps = EventActionHandlerProviderProps<BlockEditorEventActionCallableState>;
 
-export const EventActionHandlerPlugin = createDecorator(
+export const EventActionHandlerDecorator = createDecorator(
     EventActionHandlerProvider as DecoratableComponent<GenericComponent<ProviderProps>>,
     Component => {
         return function PbEventActionHandlerProvider(props) {
