+++ packages/app-page-builder/src/pageEditor/editorConfig/PageEditorConfig.tsx
@@ -9,7 +9,7 @@ interface PageEditorConfigProps {
 const BasePageEditorConfig = ({ children }: PageEditorConfigProps) => {
     return (
         <CompositionScope name={"pb.pageEditor"}>
-            <EditorConfig>{children}</EditorConfig>
+            <EditorConfig priority={"secondary"}>{children}</EditorConfig>
         </CompositionScope>
     );
 };
